requirements.txt
# Core ML dependencies
torch>=1.13
torchvision>=0.14.0
git+https://github.com/openai/CLIP.git

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Computer Vision
mediapipe>=0.10.0
ultralytics>=8.0.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Additional utilities
numpy>=1.21.0
requests>=2.28.0
